fix(core): move every remaining DI token into an import-free leaf
The leaf rule this branch introduced shipped with ten known violations of itself:
`TUS_CONFIG` in `tus.module.ts`, and 22 `AI_*` tokens spread across eleven
`ai/services/*.service.ts` files. None crashed — their graphs happened to be
acyclic — but "happens to be acyclic" is not a safety property, it is luck that
nothing is watching:

  - `@Inject(TOKEN)` is a constructor-parameter decorator, evaluated at
    class-definition time. A token declared in the file that also gets imported
    back is one back-import away from being read inside its own temporal dead zone.
  - `tsc`, `pnpm test` and `oxlint` are ALL blind to that. Only
    `pnpm run check:swc-tdz` sees it, and only after the fact.

TUS_CONFIG now lives in `tus.constants.ts`, the AI tokens in `core-ai.constants.ts`.
Both files import nothing, so they can never be mid-evaluation when someone imports
them, under any compiler.

Also removes the last real cycle in the AI module. `core-ai.service` imports
`core-ai-interaction.service` (it needs the class), and the interaction service
needed the `AiInteractionRecord` type back — a cycle held apart by ONE keyword: the
`import type`, which both compilers erase. `core-ai.service` has a constructor
`@Inject`, so `decoratorMetadata` emits `design:paramtypes` at module-evaluation
time: an IDE "organize imports" widening that import would have armed the crash
silently. The type moved to its own interface leaf, so the edge is gone rather than
merely erased.

Every old location re-exports what it lost. Public API is byte-identical: 472
exports on develop, 472 here, zero added, zero removed — verified by building both
refs and diffing.

Repo-wide cycles: 6 -> 5. The five that remain are type-only imports madge reports
but both compilers erase (empty emits) — not runtime cycles at all.
